+++ src/app/(providers)/(root)/(public)/auth/page.tsx
@@ -1,32 +1,23 @@
 'use client';
-import {auth, db} from '@/firebase';
-import {
-  Button,
-  Input,
-  Modal,
-  ModalBody,
-  ModalContent,
-  ModalFooter,
-  ModalHeader,
-  useDisclosure,
-} from '@nextui-org/react';
+import {auth} from '@/firebase';
+import {Button, Input, useDisclosure} from '@nextui-org/react';
 import {useQuery, useQueryClient} from '@tanstack/react-query';
 import {
   AuthError,
-  GithubAuthProvider,
-  GoogleAuthProvider,
   User,
   browserSessionPersistence,
-  sendPasswordResetEmail,
   setPersistence,
   signInWithEmailAndPassword,
-  signInWithPopup,
   signOut,
 } from 'firebase/auth';
-import {Firestore, collection, doc, getDocs, query, setDoc, where} from 'firebase/firestore';
-import React, {FormEvent, useEffect, useState} from 'react';
+import {useRouter} from 'next/navigation';
+import {FormEvent, useEffect} from 'react';
 import {AiOutlineEye, AiOutlineEyeInvisible} from 'react-icons/ai';
 import {MoonLoader} from 'react-spinners';
+import AuthUseStateCollection from './_components/AuthUseStateCollection';
+import PasswordResetModal from './_components/PasswordResetModal';
+import SocialLogin from './_components/SocialLogin';
+import ValidateInput from './_components/ValidateInput';
 // 사용자 인증 상태 관리 Hook
 
 const useAuthStatus = () => {
@@ -47,36 +38,36 @@ const useAuthStatus = () => {
 };
 
 // 로그인 컴포넌트
-const AuthPage: React.FC = () => {
-  const [email, setEmail] = useState<string>('');
-  const [password, setPassword] = useState<string>('');
-  const [isLoading, setIsLoading] = useState(true); // 로딩 상태 추가
+export default function AuthPage() {
   const {data: user, isFetching} = useAuthStatus();
-  //로그인 관련 유효성 검사할때 필요한 상태
-  const [emailCheck, setEmailCheck] = useState<string>('');
-  const [passwordCheck, setPasswordCheck] = useState<string>('');
-  // 정규표현식 이메일과 비밀번호 유효성검사
-  const emailValidation = /^[a-zA-Z0-9._-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,6}$/;
-  const passwordValidation = /^(?=.*[0-9])(?=.*[!@#$%^&*])[a-zA-Z0-9!@#$%^&*]{8,}$/;
-  // 로그인 에러 메시지 상태
-  const [loginError, setLoginError] = useState<string>('');
-  // 비밀번호 모달 상태
-  const [isResetModalOpen, setIsResetModalOpen] = useState(false);
-  const [resetEmail, setResetEmail] = useState('');
   const {isOpen, onOpen, onOpenChange} = useDisclosure();
 
-  // 비밀번호 눈 모양 표시 상태
-  const [isShowPassword, setIsShowPassword] = useState(false);
-  //비밀번호 찾기 완료 후에 메세지 렌더링 관련 상태
-  const [isEmailSent, setIsEmailSent] = useState(false);
-  //비밀번호 찾기 할때 가입한 이메일이 아닌 경우 렌더링 관련 상태
-  const [emailCheckMessage, setEmailCheckMessage] = useState<string>('');
+  const {
+    email,
+    setEmail,
+    password,
+    setPassword,
+    isLoading,
+    setIsLoading,
+    emailCheck,
+    setEmailCheck,
+    passwordCheck,
+    setPasswordCheck,
+    loginError,
+    setLoginError,
+    isResetModalOpen,
+    setIsResetModalOpen,
+    resetEmail,
+    setResetEmail,
+    isShowPassword,
+    setIsShowPassword,
+    isEmailSent,
+    setIsEmailSent,
+    emailCheckMessage,
+    setEmailCheckMessage,
+  } = AuthUseStateCollection();
 
-  // 비밀번호 표시 토글 함수
-  const clickTogglePasswordhandler = () => {
-    setIsShowPassword(!isShowPassword);
-  };
-
+  const router = useRouter();
   useEffect(() => {
     // 사용자의 인증 상태가 -확인되면 로딩 상태를 종료합니다.
     if (!isFetching) {
@@ -84,131 +75,33 @@ const AuthPage: React.FC = () => {
     }
   }, [isFetching]);
 
-  // 비밀번호 재설정 모달 열기
-  const openResetModal = () => setIsResetModalOpen(true);
-
-  // 비밀번호 재설정 모달 닫기
-  const closeResetModal = () => setIsResetModalOpen(false);
-
-  // 비밀번호 재설정 로직
-  const clickPasswordResetHandler = async () => {
-    const emailExists = await checkEmailExists(db, resetEmail);
-    if (!emailExists) {
-      setEmailCheckMessage('가입되지 않은 이메일입니다.');
-      return;
-    }
-    try {
-      await sendPasswordResetEmail(auth, resetEmail);
-      setIsEmailSent(true);
-      setEmailCheckMessage('');
-      //alert('비밀번호 재설정 이메일이 발송되었습니다. 이메일을 확인해 주세요.');
-      closeResetModal();
-    } catch (error) {
-      console.error('비밀번호 재설정 에러:', error);
-      setEmailCheckMessage('비밀번호 재설정에 실패했습니다. 다시 시도해 주세요.');
-    }
-  };
-
-  //비밀번호 재설정에 필요한 이메일 체크
-  const checkEmailExists = async (db: Firestore, emailToCheck: string): Promise<boolean> => {
-    const usersRef = collection(db, 'users');
-    const q = query(usersRef, where('email', '==', emailToCheck));
-    const querySnapshot = await getDocs(q);
-    return !querySnapshot.empty;
-  };
-
-  // Google 로그인 함수
-
-  const googleLogin = async () => {
-    try {
-      await setPersistence(auth, browserSessionPersistence); // 세션 지속성 설정
-      const provider = new GoogleAuthProvider();
-      const result = await signInWithPopup(auth, provider);
-      const user = result.user;
-
-      // Firestore에 저장할 사용자 정보
-      const userData = {
-        nickname: user.displayName || '기본닉네임',
-        email: user.email,
-        // 구글API에서 생년월일은 미지원으로 생년월일은 초기에 빈 값으로 설정
-        birthdate: '',
-      };
-
-      // 사용자 정보 저장
-      await setDoc(doc(db, 'users', user.uid), userData);
-
-      alert('Google 로그인 성공!');
-    } catch (error) {
-      console.error('Google 로그인 실패:', error);
-    }
-  };
-
-  // GitHub 로그인 함수
-  const githubLogin = async () => {
-    try {
-      await setPersistence(auth, browserSessionPersistence); // 세션 지속성 설정
-      const provider = new GithubAuthProvider();
-      const result = await signInWithPopup(auth, provider);
-      const user = result.user;
-
-      // 이메일이 없는 경우, 대체할 값을 사용
-      const replaceEmail = user.email || `${user.uid}@no-email.com`;
-
-      // Firestore에 저장할 사용자 정보
-      const userData = {
-        nickname: user.displayName || email.split('@')[0], // GitHub username 또는 이메일을 사용
-        email: replaceEmail,
-        //생년월일은 초기에 빈 값으로 설정
-        birthdate: '',
-      };
-
-      // 사용자 정보 저장
-      await setDoc(doc(db, 'users', user.uid), userData);
-
-      alert('GitHub 로그인 성공!');
-    } catch (error) {
-      console.error('GitHub 로그인 실패:', error);
-    }
-  };
-
-  const validateInput = () => {
-    let isValid = true;
-    setEmailCheck('');
-    setPasswordCheck('');
-
-    // 이메일 유효성 검사
-    if (!email) {
-      setEmailCheck('이메일을 입력해주세요.');
-      isValid = false;
-    } else if (!emailValidation.test(email)) {
-      setEmailCheck('유효한 이메일 형식이 아닙니다');
-      isValid = false;
-    }
-
-    // 비밀번호 유효성 검사
-    if (!password) {
-      setPasswordCheck('비밀번호를 입력해주세요.');
-      isValid = false;
-    } else if (!passwordValidation.test(password)) {
-      setPasswordCheck('비밀번호는 8자 이상, 숫자 및 특수문자를 포함해야 합니다.');
-      isValid = false;
-    }
-
-    return isValid;
+  // 비밀번호 표시 토글 함수
+  const clickTogglePasswordhandler = () => {
+    setIsShowPassword(!isShowPassword);
   };
 
   // 로그인 버튼 클릭 시 실행되는 함수
   const clickLoginHandler = async (event: FormEvent) => {
     event.preventDefault();
     setLoginError('');
-    if (!validateInput()) return;
 
+    if (
+      !ValidateInput({
+        email,
+        setEmailCheck,
+        password,
+        setPasswordCheck,
+      })
+    )
+      return;
     setPersistence(auth, browserSessionPersistence)
       .then(() => {
         return signInWithEmailAndPassword(auth, email, password);
       })
       .then(() => {
         alert('로그인 성공!');
+        setEmail('');
+        setPassword('');
       })
       .catch(error => {
         const authError = error as AuthError;
@@ -216,12 +109,18 @@ const AuthPage: React.FC = () => {
         switch (authError.code) {
           case 'auth/invalid-credential':
             setLoginError('이메일 혹은 비밀번호가 틀렸습니다.');
+            setEmail('');
+            setPassword('');
             break;
           case 'auth/too-many-requests':
             setLoginError('비밀번호 실패가 너무 많아 계정이 잠겼습니다. 나중에 다시 시도해주세요.');
+            setEmail('');
+            setPassword('');
             break;
           default:
             setLoginError('로그인 실패: ' + authError.message);
+            setEmail('');
+            setPassword('');
             break;
         }
       });
@@ -231,8 +130,13 @@ const AuthPage: React.FC = () => {
   const clickLogoutHandler = async () => {
     try {
       await signOut(auth); // Firebase에서 로그아웃 요청
+      setEmail('');
+      setPassword('');
+      router.replace('/'); // 메인페이지 이동
     } catch (error) {
       console.error(error);
+      setEmail('');
+      setPassword('');
     }
   };
 
@@ -299,63 +203,21 @@ translate-x-[13px] float-right bg-transparent text-xs  z-40 text-[#0051FF]"
         <Button size="lg" type="submit" className="mt-[40px] w-full bg-[#0051FF] font-bold text-white">
           로그인
         </Button>
-        <div className="text-center font-bold mt-[30px]">
-          <p className="text-xl">간편 로그인</p>
-          <Button onClick={googleLogin} className="h-[50px] mt-[20px] bg-transparent	">
-            <img src="/google_icon.svg" />
-          </Button>
-          <Button onClick={githubLogin} className="h-[50px] mt-[20px] bg-transparent	">
-            <img src="/github_icon.svg" />
-          </Button>
-        </div>
+        <SocialLogin />
       </form>
 
       {/* 비밀번호 재설정 모달 */}
-      <Modal
+      <PasswordResetModal
         isOpen={isOpen}
+        setIsResetModalOpen={setIsResetModalOpen}
         onOpenChange={onOpenChange}
-        classNames={{base: 'border-[#6697FF] bg-[#E5EEFF]'}}
-        size="md"
-        backdrop="opaque"
-      >
-        <ModalContent>
-          {onClose => (
-            <>
-              <ModalHeader className="flex flex-col gap-1 bg-[#E5EEFF]">비밀번호 찾기</ModalHeader>
-              <ModalBody>
-                {isEmailSent ? (
-                  // 비밀번호 재설정 이메일을 보낸 후에 표시할 내용
-                  <>
-                    <p className="text-center">비밀번호 재설정 이메일이 발송되었습니다.</p>
-                    <p className="text-center">이메일을 확인해 주세요.</p>
-                  </>
-                ) : (
-                  // 기존 이메일 입력 폼
-                  <>
-                    <label htmlFor="resetEmail">가입 시 등록한 이메일을 입력해주세요.</label>
-                    <Input
-                      size="lg"
-                      placeholder="abcde@gmail.com"
-                      value={resetEmail}
-                      onChange={e => setResetEmail(e.target.value)}
-                    />
-                    {emailCheckMessage && <p className="text-red-500 text-center">{emailCheckMessage}</p>}
-                  </>
-                )}
-              </ModalBody>
-              <ModalFooter>
-                {!isEmailSent && (
-                  <Button color="primary" className="w-full" onPress={clickPasswordResetHandler}>
-                    비밀번호 찾기
-                  </Button>
-                )}
-              </ModalFooter>
-            </>
-          )}
-        </ModalContent>
-      </Modal>
+        resetEmail={resetEmail}
+        setResetEmail={setResetEmail}
+        emailCheckMessage={emailCheckMessage}
+        setEmailCheckMessage={setEmailCheckMessage}
+        isEmailSent={isEmailSent}
+        setIsEmailSent={setIsEmailSent}
+      />
     </div>
   );
-};
-
-export default AuthPage;
+}
